refactor: remove api module
With #2052 there is no longer any need for the api module so it can be
removed.
