+++ tests/tracer/test_writer.py
@@ -8,7 +8,6 @@ import mock
 import msgpack
 import pytest
 
-from ddtrace.api import Response
 from ddtrace.compat import PY3
 from ddtrace.compat import get_connection_response
 from ddtrace.compat import httplib
@@ -16,6 +15,7 @@ from ddtrace.constants import KEEP_SPANS_RATE_KEY
 from ddtrace.internal.uds import UDSHTTPConnection
 from ddtrace.internal.writer import AgentWriter
 from ddtrace.internal.writer import LogWriter
+from ddtrace.internal.writer import Response
 from ddtrace.internal.writer import _human_size
 from ddtrace.span import Span
 from ddtrace.vendor.six.moves import BaseHTTPServer
